Avoid redaction of the Cachix cache name from logs
The Cachix cache name is stored in a secret to make it configurable
without touching the actual workflow in the repo.  However, if that name
is stored in a GitHub secret simply as a string, GitHub will redact any
matching strings from build logs, which is not really desired (because
the cache name is not really secret).

As a workaround, the GitHub secret which is used to store the Cachix
cache name now has that name wrapped in JSON; the raw cache name then
does not match any GitHub secret and does not get redacted.