Pin Nix to 2.9.2 due to the `x86_64-darwin` regex breakage
There is an unresolved problem in Nix: the behavior of `builtin.match`
is OS-specific, because it depends on the regex library implementation:

    https://www.github.com/NixOS/nix/issues/4758

Since the Nix 2.10.0 release the problem propagated to the binaries at
https://releases.nixos.org/, and `cachix/install-nix-action` by default
uses the latest release binary from that site.  Pin Nix to 2.9.2 (the
latest version that still worked) while investigating the problem.