ci: Ignore nonexistent and renamed branches
The branches for pending PRs don't need to exist when there is no
corresponding open PR.  Just leaving those branches with the content of
the last PR may result in errors when the branch goes stale and fails to
build in the current environment, therefore the best solution is to
remove the PR branches immediately after they are no longer needed, and
ignore nonexistent branches in this workflow.

Apparently `github.rest.repos.getBranch()` follows branch renames, so it
may return a branch with a different name from the requested one; ignore
such branches (they are probably archived PR branches which should not
be tested by CI).